// sourcecode/BTrak.Packages/snova-asset-module/Assets-module-components/assets-components/src/lib/snovasys-asset-management/models/asset-input-model.ts
export class AssetInputModel {
    pageNumber: number;
    pageSize: number;
    searchText: string;
    sortBy: string;
    sortDirectionAsc: boolean;
    allAssigned: boolean;
    byUser: boolean;
    allDamaged: boolean;
    productDetailsId: string;
    supplierId: string;
    assignedToEmployeeId: string;
    isEmpty: boolean;
    isVendor: boolean;
    userId: string;
    branchId: string;
    searchAssetCode: string;
    allPurchasedAssets: boolean;
    seatingId: string;
    assetId: string;
    entityId: string;
    activeAssignee: boolean;
    purchasedDate: Date;
    assignedDate: Date;
    isUpdateMultipleAssignee: boolean;
    assetIds: string;
    approvedByUserId: string;
    isListOfAssetsPage: boolean;
    isWriteOff: boolean;
}